# go: reader: make ordered iteration faster