--- a/go/conformance/test-read-conformance/main.go
+++ b/go/conformance/test-read-conformance/main.go
@@ -354,7 +354,7 @@ func readIndexed(w io.Writer, filepath string) error {
 	knownChannelIDs := make(map[uint16]bool)
 
 	for {
-		schema, channel, message, err := it.Next2(nil)
+		schema, channel, message, err := it.NextInto(nil)
 		if errors.Is(err, io.EOF) {
 			break
 		}