… .vscode/settings.json
… go/.golangci.yaml
… go/cli/mcap/cmd/sort_test.go
… go/mcap/lexer.go
… go/mcap/parse.go
… go/mcap/range_index_heap.go
… go/mcap/range_index_heap_test.go
… go/mcap/reader.go
… go/mcap/reader_test.go
… go/mcap/slicemap.go
… go/mcap/slicemap_test.go
… go/mcap/unindexed_message_iterator.go
… go/mcap/version.go
… go/ros/ros2db3_to_mcap_test.go