@wkalt 2024-05-07
I guess my concrete thought is, can we make this the Next signature:

```
Next(*Message) (*Schema, *Channel, *Message, error)
```

and get performance identical to the lexer, with that being the only required new interface. I suspect the answer is yes -- and if so, that seems like a better outcome.