@james-rms 2024-05-08
Regarding the signature - i've pushed an update which makes the API `ReadInto(*Message) (*Schema, *Channel, error)`. i can't see a point returning the message pointer if it's always going to be the same as the passed-in pointer.

Regarding always looking up the schema and channel - this makes a noticeable difference to the benchmark. I tested 2d2d767e4 against a patch that skips lookup and always returns `nil` for schemas and channels:

#### Without lookup
```
go test ./... -run=^$ -bench=BenchmarkReader -benchmem -benchtime=20s 
BenchmarkReader/no_index-8                   100         209160724 ns/op         149.31 MB/s       4892764 msg/s        18126126 B/op       2883 allocs/op
BenchmarkReader/with_index-8                  62         381960212 ns/op          74.79 MB/s       2450602 msg/s         7425817 B/op       5960 allocs/op
PASS
```

#### With lookup
```
go test ./... -run=^$ -bench=BenchmarkReader -benchmem -benchtime=20s 
BenchmarkReader/no_index-8                   100         230150386 ns/op         132.22 MB/s       4332659 msg/s        16929751 B/op       2871 allocs/op
BenchmarkReader/with_index-8                  55         429351720 ns/op          71.90 MB/s       2356059 msg/s         7425896 B/op       5960 allocs/op
PASS
```

My idea when separating the lookup to its own function is that for users who need to go fast, they rarely want the schema and channel information more than once. 
- If you're splitting an MCAP up,  for the first message you figure out which channel IDs need to go into which stream, then you build up a mapping and key off the channel ID after then.
- If you're trying to decode messages, you instantiate your deserializers with schema information once per channel, you store them in a mapping of channel ID -> deserializer, and you never look at the schema information again.