@james-rms 2024-05-08
Also re: testing against the lexer: I have a rough estimate here:
```
BenchmarkReader/lexer-8                    10         175206304 ns/op         173.64 MB/s       5689735 msg/s        17305244 B/op       1260 allocs/op
```