@wkalt 2024-05-08
Is that benchmark you posted for the lexer comparable to the other two posted earlier? That seems like a great basis for comparison